+++ lib/src/ast/sass/supports_condition/interpolation.dart
@@ -2,9 +2,11 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+import 'package:meta/meta.dart';
 import 'package:source_span/source_span.dart';
 
 import '../expression.dart';
+import '../interpolation.dart';
 import '../supports_condition.dart';
 
 /// An interpolated condition.
@@ -17,6 +19,15 @@ final class SupportsInterpolation implements SupportsCondition {
   final FileSpan span;
 
   SupportsInterpolation(this.expression, this.span);
+
+  /// @nodoc
+  @internal
+  Interpolation toInterpolation() => Interpolation([expression], [span], span);
+
+  /// @nodoc
+  @internal
+  SupportsInterpolation withSpan(FileSpan span) =>
+      SupportsInterpolation(expression, span);
 
   String toString() => "#{$expression}";
 }