+++ lib/src/ast/sass/supports_condition/negation.dart
@@ -2,8 +2,12 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+import 'package:meta/meta.dart';
 import 'package:source_span/source_span.dart';
 
+import '../../../interpolation_buffer.dart';
+import '../../../util/span.dart';
+import '../interpolation.dart';
 import '../supports_condition.dart';
 import 'operation.dart';
 
@@ -17,6 +21,18 @@ final class SupportsNegation implements SupportsCondition {
   final FileSpan span;
 
   SupportsNegation(this.condition, this.span);
+
+  /// @nodoc
+  @internal
+  Interpolation toInterpolation() => (InterpolationBuffer()
+        ..write(span.before(condition.span).text)
+        ..addInterpolation(condition.toInterpolation())
+        ..write(span.after(condition.span).text))
+      .interpolation(span);
+
+  /// @nodoc
+  @internal
+  SupportsNegation withSpan(FileSpan span) => SupportsNegation(condition, span);
 
   String toString() {
     if (condition is SupportsNegation || condition is SupportsOperation) {